Separate WC Adapter from Argent Adapter
